reverting additional logic added to createNamespace
logic was added as test for adding multiple ACKAddOns was failing as both addons were trying to create the same namespace. It was most likely an intentional design choice for the user to override the default prop to create namespace, as the prop comment mentions it. Modified test so that only the first ACKAddOn added to the list spefies to create namespace and others do not