Add S3 CSI driver IAM policy as addon
This commit adds a new file `iam-policy.ts` that contains the IAM policy statements for the S3 CSI driver. It also adds a new file `index.ts` that implements the S3 CSI driver addon. The addon creates a service account, attaches the IAM policy to the service account's role, and deploys the addon to the cluster. The commit also includes some helper functions and default options for the addon.